39 enhance database (#68)
* Readme update (#66)

* Reworking sql queries on recipes

* recipe querys reworked

* modified recipe querys, started to rework dish querys

* Reworked dish querys

* Basket query rework continuation, fixed logic issue on UI updating on basket

* basket querys reworked

* updated README

+++ backend/backend/db/delete/recipes.php
@@ -8,31 +8,63 @@ require __DIR__ . "/../connection.php";
 
 $data = $resource["data"];
 $message = "Recipe";
-if(isset($resource["data"]["dependencies"])){
-    $dependencies = $resource["data"]["dependencies"];
-    $stmtDep = $pdo->prepare("
-        DELETE FROM dishes
-        WHERE id = :id
-    ");
-    foreach ($dependencies as $dep) {
-        $stmtDep->execute(["id" => (int)$dep]);
+
+try {
+    $pdo->beginTransaction();
+
+
+    //DELETE RECIPE DEPENDENCIES
+    if(isset($resource["data"]["dependencies"]) && !empty($resource["data"]["dependencies"])){
+        $dependencies = $resource["data"]["dependencies"];
+
+        $values = [];
+        $params = [];
+
+        foreach($dependencies as $i => $dep){
+            $values[] = ":id$i";
+            $params["id$i"] = (int)$dep;
+        }
+
+        $sql = "
+            DELETE FROM dishes
+            WHERE id IN  (" . implode(", ", $values) . ")
+        ";
+
+        $stmt = $pdo->prepare($sql);
+        $stmt->execute($params);
+        unset($stmt);
+
+        $message = "Selected items";
     }
 
-    $message = "Selected items";
+    //DELETE RECIPE
+    $stmt = $pdo->prepare("
+        DELETE FROM recipes 
+        WHERE id = :id
+    ");
+    $stmt->execute([
+        "id" => (int)$data["id"]
+    ]);
+
+    $pdo->commit();
+
+    http_response_code(200);
+    header("Content-Type: application/json");
+    echo json_encode(["success" => "$message deleted successfully!"]);
+
+} catch (Throwable $th) {
+    if ($pdo->inTransaction()) {
+        $pdo->rollBack();
+    }
+
+    http_response_code(500);
+    header("Content-Type: application/json");
+    echo json_encode(["error" => "Failed to delete recipe.", "debug" => $th->getMessage()]);
+
+} finally{
+    unset($pdo);
 }
 
-$stmtDelRec = $pdo->prepare("
-    DELETE FROM recipes 
-    WHERE id = :id
-");
 
-$stmtDelRec->execute([
-    "id" => (int)$data["id"]
-]);
-
-http_response_code(200);
-header("Content-Type: application/json");
-echo json_encode(["success" => "$message deleted successfully!"]);
-exit;
 
 ?>
